user_auth_ms/requirements.txt
django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
celery==5.4.0
redis==5.2.1
twilio==9.4.5